redis: Extract `time_buckets` module
