+++ src/redis/messages.rs
@@ -10,7 +10,8 @@ use _redis::{Commands, Connection};
 use regex::Regex;
 use itertools::Itertools;
 
-use redis::RedisExecutor;
+use redis::executor::RedisExecutor;
+use redis::time_buckets::*;
 
 #[derive(Serialize, Deserialize, Debug)]
 struct RedisOGNRecord {
@@ -205,22 +206,6 @@ impl OGNRedisCommands for Connection {
     }
 }
 
-fn bucket_times_between(from: DateTime<Utc>, to: DateTime<Utc>) -> Vec<i64> {
-    let from_bucket_time = from.to_bucket_time();
-    let to_bucket_time = to.to_bucket_time();
-    (from_bucket_time..=to_bucket_time).step_by(60 * 60).collect()
-}
-
-trait ToBucketTime {
-    fn to_bucket_time(&self) -> i64;
-}
-
-impl ToBucketTime for DateTime<Utc> {
-    fn to_bucket_time(&self) -> i64 {
-        self.with_minute(0).unwrap().with_second(0).unwrap().timestamp()
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use std::mem::size_of;
@@ -265,37 +250,5 @@ mod tests {
         assert_eq!(records.len(), 3);
         assert_eq!(records[0].seconds, 123);
         assert_eq!(records[1].altitude, 2345);
-    }
-
-    #[test]
-    fn test_bucket_times() {
-        use chrono::prelude::*;
-        use super::bucket_times_between;
-
-        fn check(from: &str, to: &str, expected: Vec<&str>) {
-            assert_eq!(bucket_times_between(from.parse().unwrap(), to.parse().unwrap()),
-                       expected.iter().map(|it| it.parse::<DateTime<Utc>>().unwrap().timestamp()).collect::<Vec<_>>());
-        }
-
-        check("2018-08-07T01:23:45Z", "2018-08-07T01:23:45Z", vec![
-            "2018-08-07T01:00:00Z",
-        ]);
-
-        check("2018-08-07T01:23:45Z", "2018-08-07T05:00:00Z", vec![
-            "2018-08-07T01:00:00Z",
-            "2018-08-07T02:00:00Z",
-            "2018-08-07T03:00:00Z",
-            "2018-08-07T04:00:00Z",
-            "2018-08-07T05:00:00Z",
-        ]);
-
-        check("2018-08-06T22:00:00Z", "2018-08-07T03:59:59Z", vec![
-            "2018-08-06T22:00:00Z",
-            "2018-08-06T23:00:00Z",
-            "2018-08-07T00:00:00Z",
-            "2018-08-07T01:00:00Z",
-            "2018-08-07T02:00:00Z",
-            "2018-08-07T03:00:00Z",
-        ]);
     }
 }
