+++ src/redis/mod.rs
@@ -1,5 +1,6 @@
 mod executor;
 mod messages;
+mod time_buckets;
 
 pub use redis::executor::RedisExecutor;
 pub use redis::messages::*;
